+++ ft_strlcpy.c
@@ -6,11 +6,13 @@
 /*   By: pgomez-r <pgomez-r@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 13:00:07 by pgomez-r          #+#    #+#             */
-/*   Updated: 2022/10/05 13:00:18 by pgomez-r         ###   ########.fr       */
+/*   Updated: 2022/10/13 14:05:03 by pgomez-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+#include "libft.h"
+
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
 	unsigned int	cnt;
